feat: add files of lab4

+++ labs/src/main/java/labs/service/impl/DayServiceImpl.java
@@ -6,19 +6,23 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.github.fge.jsonpatch.JsonPatch;
 import com.github.fge.jsonpatch.JsonPatchException;
+import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.util.List;
+import labs.aspect.LogExecution;
 import labs.dao.DayDao;
 import labs.dto.DayDto;
+import labs.exception.ExceptionMessages;
+import labs.exception.NotFoundException;
+import labs.exception.ValidationException;
 import labs.model.Day;
 import labs.service.DayService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 @Service
+@LogExecution
 public class DayServiceImpl implements DayService {
     private final DayDao dayDao;
 
@@ -29,51 +33,66 @@ public class DayServiceImpl implements DayService {
 
     @Override
     public DayDto getDayById(int id) {
+        if (!dayDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.DAY_NOT_FOUND, id));
+        }
         Day day = dayDao.getDayById(id);
         return DayDto.toDto(day);
     }
 
     @Override
     public int addDay(Day day) {
-        return dayDao.addDay(day);
+        Day dayFromDb;
+        try {
+            dayFromDb = getDayByDate(day.getDate()).get(0).fromDto();
+        } catch (NotFoundException ex) {
+            return dayDao.addDay(day);
+        }
+        return dayFromDb.getId();
     }
 
     @Override
     public List<DayDto> getAllDays() {
         List<Day> days = dayDao.getAllDays();
+        if (days.isEmpty()) {
+            throw new NotFoundException(ExceptionMessages.DAYS_NOT_FOUND);
+        }
         return days.stream().map(DayDto::toDto).toList();
     }
 
     @Override
     public ResponseEntity<String> deleteDayById(int id) {
+        if (!dayDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.DAY_NOT_FOUND, id));
+        }
         return dayDao.deleteDayById(id);
     }
 
     @Override
     public ResponseEntity<DayDto> updateDayById(int id, JsonPatch json)
-            throws JsonProcessingException {
+            throws JsonProcessingException, JsonPatchException {
+        if (!dayDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.DAY_NOT_FOUND, id));
+        }
         if (json.toString().contains("id") || json.toString().contains("meals")) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            throw new ValidationException(String.format(ExceptionMessages.PATCH_VALIDATION_EXCEPTION,
+                    "'id' and 'meals'"));
         }
         ObjectMapper objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
         Day day = dayDao.getDayById(id);
         JsonNode node;
-        try {
-            node = json.apply(objectMapper.convertValue(day, JsonNode.class));
-        } catch (JsonPatchException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
+        node = json.apply(objectMapper.convertValue(day, JsonNode.class));
         day = objectMapper.treeToValue(node, Day.class);
         Day updatedDay = dayDao.updateDayById(id, day);
         return ResponseEntity.ok(DayDto.toDto(updatedDay));
     }
 
     @Override
-    public List<DayDto> getDayByDate(LocalDate date) {
+    public List<DayDto> getDayByDate(@NotNull LocalDate date) {
         List<Day> days = dayDao.getDayByDate(date);
         if (days.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            throw new NotFoundException(ExceptionMessages.DAYS_NOT_FOUND_BY_DATE);
         }
         return days.stream().map(DayDto::toDto).toList();
     }
