+++ labs/src/main/java/labs/service/impl/ProductServiceImpl.java
@@ -10,8 +10,13 @@ import com.github.fge.jsonpatch.JsonPatchException;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import labs.aspect.LogExecution;
+import labs.dao.MealDao;
 import labs.dao.ProductDao;
 import labs.dto.ProductDto;
+import labs.exception.ExceptionMessages;
+import labs.exception.NotFoundException;
+import labs.exception.ValidationException;
 import labs.model.Product;
 import labs.service.ProductService;
 import lombok.extern.slf4j.Slf4j;
@@ -19,51 +24,55 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 @Slf4j
 @Service
+@LogExecution
 public class ProductServiceImpl implements ProductService {
     private final ProductDao productDao;
+    private final MealDao mealDao;
 
     @Autowired
-    public ProductServiceImpl(ProductDao productDao) {
+    public ProductServiceImpl(ProductDao productDao, MealDao mealDao) {
         this.productDao = productDao;
+        this.mealDao = mealDao;
     }
 
     @Override
     public ProductDto getProductById(int id) {
+        if (!productDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.PRODUCT_NOT_FOUND, id));
+        }
         Product product = productDao.getProductById(id);
         return ProductDto.toDto(product);
     }
 
     @Override
     public List<Product> getProductsByQuery(String query) throws IOException {
-        long startTime = System.currentTimeMillis();
         ObjectMapper mapper = new ObjectMapper();
         JsonNode node = ProductServiceImpl.getJsonFromExternalApi(query);
-        log.info("Time elapsed for product by query = " + (System.currentTimeMillis() - startTime) + "ms");
-        return mapper.treeToValue(node.get("items"), new TypeReference<List<Product>>() {});
+        List<Product> products = mapper.treeToValue(node.get("items"), new TypeReference<List<Product>>() {});
+        if (products.isEmpty()) {
+            throw new NotFoundException(String.format(ExceptionMessages.PRODUCTS_NOT_FOUND_BY_QUERY, query));
+        }
+        return products;
     }
 
-    private static JsonNode getJsonFromExternalApi(String query) throws IOException {
+    public static JsonNode getJsonFromExternalApi(String query) throws IOException {
         OkHttpClient client = new OkHttpClient();
         Request requestForExternalApi = new Request.Builder().url("https://api.calorieninjas.com/v1/nutrition?query=" +
                 query).addHeader("X-Api-Key", "").build();
-        long startTime = System.currentTimeMillis();
         Response responseFromExternalApi = client.newCall(requestForExternalApi).execute();
-        log.info("Time elapsed for API request = " + (System.currentTimeMillis() - startTime) + "ms");
         ObjectMapper mapper = new ObjectMapper();
         return mapper.readTree(responseFromExternalApi.body().string());
     }
 
     @Override
     public List<Integer> addProductByMealId(int mealId, ProductDto productDto) {
-        if (productDto.getId() != 0) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (!mealDao.existsById(mealId)) {
+            throw new NotFoundException(String.format(ExceptionMessages.MEAL_NOT_FOUND, mealId));
         }
         List<Integer> ids = new ArrayList<>();
         if (productDto.getWeight() != 100) {
@@ -88,6 +97,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public List<Integer> addProductsByQueryAndMealId(int mealId, String query) throws IOException {
+        if (!mealDao.existsById(mealId)) {
+            throw new NotFoundException(String.format(ExceptionMessages.MEAL_NOT_FOUND, mealId));
+        }
         List<Product> products = getProductsByQuery(query);
         List<Product> updatedProducts = products.stream().map(this::setWeightAndCalories).toList();
         return updatedProducts.stream()
@@ -96,43 +108,61 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public List<ProductDto> getAllProductsByMealId(int mealId) {
+        if (!mealDao.existsById(mealId)) {
+            throw new NotFoundException(String.format(ExceptionMessages.MEAL_NOT_FOUND, mealId));
+        }
         List<Product> products = productDao.getAllProductsByMealId(mealId);
+        if (products.isEmpty()) {
+            throw new NotFoundException(String.format(ExceptionMessages.PRODUCTS_NOT_FOUND_BY_MEAL_ID,
+                    mealId));
+        }
         return products.stream().map(ProductDto::toDto).toList();
     }
 
     @Override
     public List<ProductDto> getAllProducts() {
-        return productDao.getAllProducts().stream().map(ProductDto::toDto).toList();
+        List<Product> products = productDao.getAllProducts();
+        if (products.isEmpty()) {
+            throw new NotFoundException(ExceptionMessages.PRODUCTS_NOT_FOUND);
+        }
+        return products.stream().map(ProductDto::toDto).toList();
     }
 
     @Override
     public ResponseEntity<String> deleteProductsByMealId(int mealId) {
+        if (!mealDao.existsById(mealId)) {
+            throw new NotFoundException(String.format(ExceptionMessages.MEAL_NOT_FOUND, mealId));
+        }
         return productDao.deleteProductsByMealId(mealId);
     }
 
     @Override
     public ResponseEntity<String> deleteProductById(int id) {
+        if (!productDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.PRODUCT_NOT_FOUND, id));
+        }
         return productDao.deleteProductById(id);
     }
 
     @Override
     public ResponseEntity<ProductDto> updateProductById(int id, JsonPatch json)
-            throws JsonProcessingException {
+            throws JsonProcessingException, JsonPatchException {
+        if (!productDao.existsById(id)) {
+            throw new NotFoundException(String.format(ExceptionMessages.PRODUCT_NOT_FOUND, id));
+        }
         ObjectMapper objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
         Product product = productDao.getProductById(id);
+        ProductDto productDto = ProductDto.toDto(product);
         if (json.toString().contains("id") || json.toString().contains("meals") ||
                 json.toString().contains("weight")) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            throw new ValidationException(String.format(ExceptionMessages.PATCH_VALIDATION_EXCEPTION,
+                    "'id', 'meals' and 'weight'"));
         }
         JsonNode node;
-        try {
-            node = json.apply(objectMapper.convertValue(product, JsonNode.class));
-        } catch (JsonPatchException e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
-        product = objectMapper.treeToValue(node, Product.class);
-        Product updatedProduct = productDao.updateProduct(id, product);
-        return ResponseEntity.ok(ProductDto.toDto(updatedProduct));
+        node = json.apply(objectMapper.convertValue(productDto, JsonNode.class));
+        productDto = objectMapper.treeToValue(node, ProductDto.class);
+        product = productDto.fromDto();
+        return ResponseEntity.ok(ProductDto.toDto(productDao.updateProduct(id, product)));
     }
 }
